Bring up interface when adding it to bridge
When deploying on metal with MAAS enlisted and bound interfaces
will be brought up at boot automatically.

However, the user may choose to use interfaces that are either hot
plugged or not configured/enlisted with substrate.

Thus the charm should bring up any interface before adding it to
a bridge.

NOTE: this was added to the chassis charm in [0] but was lost when
the code was moved to the layer.

0: https://review.opendev.org/#/c/692493/