Change ovn layer to work with os_actions in openstack layer
The openstack layer [2] has a default implementation of actions which
are implemented in the os_actions modules.  This uses
charms_openstac.bus.discover() to locate and import the concrete charm
class and instantiate it.  This is needed for the default actions to
work.  The linked bug [1] is due to the module charms.ovn_central not
existing as the two dependent charm layers (ovn-chassis and
ovn-dedicated-chassis) implement them in uniquely named modules
(ovn_chassis and ovn_dedicated_chassis).  The bug [1] is fixed by
modifying the charm layers to place their concrete charm class modules
into /lib/charm/openstack upon which everything will 'just work'.

[1] LP bug: https://bugs.launchpad.net/charm-ovn-chassis/+bug/1861069
[2] https://review.opendev.org/p/openstack/charm-layer-openstack