Fix package/service adaption in specialized classes
When the charm code executes the default release selector code
will instantiate all incarnations of the non-abstract charm
classes as a side effect of figuring out which one to use.

At present we set the package and service list as class variables,
subsequently the specialized class initializers will extend on
the same variables and we end up with a charm class with an
incorrect package and service specification.

To fix this we assign the package and service variable in the OVN
Charm base class initializer.

We also remove activation of the default select release handler
in the layer as it will execute too late.  Each individual charm
must execute the handler as early as possible.