Populate the Open vSwitch system-id.conf file
The charm configures this at run time, however in the event of a
restart of the ``openvswitch-switch`` service the init script will
read a UUID generated at initial service startup from that file.

With RBAC at play it is imperative that Open vSwitch presents
itself with the correct sytem ID or it will be denied access.