Fix certificate request and response handling
The charm class was supposed to override the default certificate
request handler, but due to a typo it did not.

Be more defensive in the code that writes certificates out to disk.
Even with our certificate request handler in place the charm may
get into situations where there are multiple certificate requests
on the relation.