Remove flawed update-status processing avoidance code
The current attempt at not doing processing on update-status does
not work. This is due to the default handler clears the flag after
running assess_status and all the handlers that was gated on
`run-default-upgrade-charm` will then be queued and run again.

There is no simple way to avoid this with reactive, so I'll just
remove the non-functioning code to avoid future confusion.