Do full hook execution on resume
A part of the migration strategy to OVN is to start the chassis
charms paused only to resume them as soon as cleanup after the
previous SDN has been completed.

For this to work we need to run a full hook execution on resume to
make sure the system is configured properly.