SR-IOV: match on PCI address, don't do runtime config
Populate the sriov-netplan-shim interfaces.yaml config with PCI
address of interfaces. This is necessary as the VFs may be
configured at a time in the boot sequence where interface naming
has not settled yet.

Stop doing runtime reconfiguration of SR-IOV, in addition to being
detrimental to any virtual machine instance consuming the VF this
will break NIC firmware in some configurations.

Merge after:
https://github.com/juju/charm-helpers/pull/549

Closes-Bug: #1908351