Move OVNChassisCharm propoerties out of __init__
When the ovn-chassis charm is deployed with new-units-paused=true,
the charm.ovn-chassis.enable-openstack flag is not set because the
handler that sets that flag does not run when units are paused. When
the migration is complete the operator runs a resume on all the
ovn-chassis units. This resume in-turn runs a config-changes hook
from within the resume action. When the config-change hook fires the
previously guarded handler now runs and executes the code that sets
the charm.ovn-chassis.enable-openstack state and then runs a
charm_instance.install() *1. However the charm instance that
charm.provide_charm_instance() provided is cached *2 from a previous
handler. When the first handler ran which greateed an instance of the
class charm.ovn-chassis.enable-openstack was not set and since the
list of packages, services and restart map is amended in the init
method depending on whether charm.ovn-chassis.enable-openstack flag
is set cached class does not include the metadata service in its
package,service or restart map objects. This change moves package,
service or restart maps to be properties that are evaluated when
accessed rather than at __init__ time.

Closes-Bug: #1912471