Ensure disable-mlockall is handled at deploy time
The prior commit that enable disable-mlockall support only restarts
openvswitch-switch when the disable-mlockall config value is changed. It
didn't handle deploy-time restart of openvswitch-switch.

This patch ensures that /etc/default/openvswitch-swith is rendered and
openvswitch-switch is restarted directly after install. This ensures that
when the disable-mlockall config option is unset, mlockall actually gets
disabled when running in a container. See config.yaml description for more
details.

Related-Bug: #1906280