Avoid unnecessary restart of OVS on initial deploy
The commited fix for LP: #1906280 has guard rails to only restart
Open vSwitch on initial install.

However, this is not sufficient as the charm may be deployed to a
unit with a running Open vSwitch data plane as laid out in the OVS
to OVN migration guide [0].

Rework so that we only do the restart when necessary.

0: https://docs.openstack.org/project-deploy-guide/charm-deployment-guide/latest/app-ovn.html#migration-from-neutron-ml2-ovs-to-ml2-ovn
Related-Bug: #1906280