Improve handling of major version upgrades
Setting the ``external_ids:ovn-match-northd-version`` value to
'true' will make the ovn-controller refrain from making updates to
the data plane tables in the event of a version mismatch.

This in combination with stopping the ovn-controller with the
`ovn-ctl stop_controller --restart` command will allow upgrades
to progress with little or no data plane downtime.
(Note that we will accomplish this by a separate proposal to the
OVN package itself in Ubuntu.)

As soon as the central components are upgraded ovn-controller will
notice and resume (re-)programming of the local Open vSwitch data
plane.

Closes-Bug: #1940043