Do not attempt to set link status for DPDK interfaces
When DPDK is enabled interfaces are not visible to standard Linux
network administration tools.  The c-h OVS ``del_bridge_port``
helper will by default attempt to bring link down for interfaces
removed from a bridge.

Ask the helper to not do this when DPDK is enabled to avoid a
hook error.