Fix handling of other_config for dpdk and hwol
The current code does not actually get the values from the system
and subsequently the charm will always set it's view of the world
and think it will need to restart Open vSwitch.

The affected handlers are only called on config-changed and when
either DPDK or Hardware Offload features are enabled, so the
impact is not as severe as one might think.

Closes-Bug: #1922335
Co-Authored-By: bayani.carbone@canonical.com