Switch to stestr instead of os-testr
Currently test invocations are broken due to a change in the dependency
chain: os-stestr no longer depends on stestr which is used to invoke
tests in tox.ini. It has to be explicitly specified as a dependency due
to the latest changes.

See the links below for context:

https://meetings.opendev.org/irclogs/%23openstack-qa/%23openstack-qa.2022-03-22.log.html#t2022-03-22T15:52:28
http://lists.openstack.org/pipermail/openstack-discuss/2022-March/027784.html
http://lists.openstack.org/pipermail/openstack-discuss/2022-March/027785.html

https://github.com/openstack/os-testr/commit/6f10535042ac3220a5dd9f21d25dd20bb14ba523
(actually removes stestr from os-testr dependencies which intentionally
breaks everybody relying on stestr getting installed alongside
os-testr)