Add a new option: vpd-device-spec (#60)
This option can be used to specify a device holding a card serial number
of a networking controller present in its Vital Product Data (VPD) to
enable SmartNIC DPU support.

Currently the format only supports devices on the 'pci' bus with device
and vendor ids specified. The schema could be extended to support more
buses and parameters in the future.

Config validation errors are now accumulated in the config adapter and
then merged together to form the final status message for a unit.