Add prometheus-ovs-exporter snap support
Both ovn-dedicated-chassis and ovn-chassis can utilize
the prometheus-ovs-exporter snap in order to provide metrics to
prometheus.

This change adds an option to specify a channel for the snap or lack
thereof (in which case it will not be installed or removed).

The snap is currently installed in devmode due to the lack of relevant
interfaces that the exporter needs.