iptables NOTRACK rules for excluding GENEVE traffic from nf_conntrack
On some busy cloud deployments, it has been reported that the nodes
hosting the ovn-chassis are getting their nf-conntrack tables full
and starting to drop connections. The reason is that GENEVE uses
source port randomization, and the number of "unique" flows is high
from the nf-conntrack's perspective. This is no surprise since flows
are usually identified with their five-tuple (srcip/[srcport]/dstip/dstport/tproto)
by network elements, and GENEVE is leveraging this fact to have an
even distribution in load-balancing systems in between [1]. The
randomization causes the nf_conntrack table to be filled with many
GENEVE flows, eventually leading to connection drops in a busy
environment. As there is no particular reason and benefit to track
these flows at the moment, the solution is to exclude GENEVE traffic
from nf-conntrack tracking. This can be done by putting rules with
`-j NOTRACK` jump into relevant iptables chains, which many people
already use as a solution to this problem.

This change incorporates the relevant rules to the charm code, so the
rules become present by default.

Closes-bug: #1978806

[1]: https://datatracker.ietf.org/doc/html/rfc8926#section-3.3