Use interface type to determine linkdown action
When removing a port from a bridge, we need to pass on information
to the c-h library function for whether it should change the link
state of the port with the `ip link` command after removing it
from the bridge.

Interfaces associated with a DPDK driver are not visible to the
regular system administration tools such as `ip`, and we need to
tell the library not to attempt changing link state for such
interfaces.

The current logic makes use of the charm configuration to
determine this, however, that does not work well when the charm
is transitioning the system from one state to the other, as there
may very well exist DPDK interfaces in Open vSwitch even though
the charm is not configured to use DPDK.

Add a helper method to determine port type based on its interface
records and use this method to inform the c-h library whether it
should attempt to manipulate link state or not.

Note that this change depends on juju/charm-helpers#731

Closes-Bug: #1989192