Fix a regression introduced in 0851229e
A bridge interface mapping may be set to an empty string which was not
previously handled as invalid config - instead it resulted in an empty
BridgeInterfaceMap which was handled by configure_bridges accordingly
and ovn-cms-opts were set.

The regression made it such that ovn-cms-opts were not set in a case of
an empty BridgeInterfaceMap where as this should only happen when there
is an invalid config specified to the charm.

1) Test failure:
https://review.opendev.org/c/x/charm-ovn-dedicated-chassis/+/857878/2#message-37afbdf3f745d77719aa43afe5234aaa87d70468
2) Analysis:
https://review.opendev.org/c/x/charm-ovn-dedicated-chassis/+/857878/comments/ecde1283_4f1ba2c3
3) PDB output:
https://review.opendev.org/c/x/charm-ovn-dedicated-chassis/+/857878/comments/ecde1283_4f1ba2c3