Pin Jinja2 to avoid build failure on Focal
As concluded in pallets/jinja#1496 Jinja2 requires a newer version
of setuptools than provided in Ubuntu Focal.

It is currently not possible to tell charm-tools to upgrade
setuptools in the virtualenv used for building of wheels
(juju/charm-tools#646), so the only alternative is to pin Jinja2
to avoid the problematic behavior.